requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
jinja2>=3.1.2
colorama>=0.4.6
flask>=3.0.0
